+++ src/pages/profile/[id].astro
@@ -1,7 +1,7 @@
 ---
 import Base from '@layouts/Base.astro';
 import Profile from '@components/profile/Profile';
-import user from '@store/userStore';
+
 const { id } = Astro.params;
 // request for user data
 // const user = ...
